weather_api/requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
asgiref==3.10.0
asyncio==4.0.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.3
colorama==0.4.6
Django==5.1.1
djangorestframework==3.15.2
drf-yasg==1.21.11
flake8==7.1.1
frozenlist==1.8.0
idna==3.10
inflection==0.5.1
iniconfig==2.3.0
magic-filter==1.0.12
mccabe==0.7.0
multidict==6.7.0
packaging==25.0
pluggy==1.6.0
propcache==0.4.1
pycodestyle==2.12.1
pydantic==2.11.10
pydantic_core==2.33.2
pydocstyle==6.3.0
pyflakes==3.2.0
Pygments==2.19.2
pyTelegramBotAPI==4.22.1
pytest==9.0.2
pytest-django==4.11.1
pytest-lazy-fixtures==1.4.0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.3
snowballstemmer==3.0.1
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
yarl==1.22.0
